+++ chrome_extension/BannerTools/assets/js/BannerToolsBackground.js
@@ -1,3 +1,12 @@
+chrome.commands.onCommand.addListener(function(command) {
+	if(command == "toggle-feature"){
+	    chrome.storage.sync.set({'uniqueID_disable': 'false'});
+		chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
+		  chrome.tabs.sendMessage(tabs[0].id, {ExpandPanel: "true"});
+		});
+	}
+});
+
 chrome.browserAction.onClicked.addListener(function(tab) {
     chrome.storage.sync.set({'uniqueID_disable': 'false'});
 	chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
@@ -7,8 +16,13 @@ chrome.browserAction.onClicked.addListener(function(tab) {
 
 chrome.runtime.onMessage.addListener(
 	function(request, sender, sendResponse) {
-		if ((request.execute_screenshot.indexOf("true") >= 0)){
-			var adSize = request.execute_screenshot;
+		if (request.resetZoom == "true"){
+			chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
+			  chrome.tabs.setZoom(tabs[0].id, 1);
+			});
+		}
+		else if ((request.executeScreenshot.indexOf("true") >= 0)){
+			var adSize = request.executeScreenshot;
 			adSize =  adSize.split(';').pop();
 
 			var adWidth = adSize.split('@')[0];